main: remove the --in-place flag
The --output flag was removed to handle formatting multiple files at the
same time. But since that flag is missing, it doesn't make sense to
output multiple files on stdout as well.

If the user wants to format a file and write the output to a different
place, use stdin and stdout.