implement nixpkgs-fmt --check
Send all file paths that have/might be changed to stdout. This can then
be consumed by other unixy tools.

When running nixpkgs-fmt with --check, the tool will not write the
changes it found, and exit with status 1 if any files with changes have
been found. This is mainly useful for CI.