commit whitespace changes before computing indentation
The biggest issue with our current formatting infra is that patterns
are evaluated on the original syntax tree, and not on the formatting
model, which might contain whitespace changes.

A particular bad issue here is when whitespace changes add or remove
newlines, but then we use newline info to calculate indent.
Originally, the thinking was that this is not necessary, as rules
would only depend on the shape of the tree and not on the spacing, but
turns out this is not the case in practice.

The "obvious" fix here is to make our formatting model & predicated
more complicated, so that they can take modified whitespace into
account. But this PR proposes something more stupid, yet more
elegant (subjectively) -- we just commit whitespace changes to the
syntax tree before re-indent pass.

That is, we first fix whitespace, then modify the syntax tree to take
updated ws into account (by using to_string and reparse), and then
re-indent the new tree.

This will be a bit slower, as we have one extra step of tree
re-building, but the perf should not matter that much (rust is not a
slow language), and we can always make the internal model more
perofrmant and zero-copy later.