Reformat test_data to match new multiline string rule.
This commit trying to address the test_data based on the new string interpolation rule.

The `.bad.nix` file changes is necessary because in the new string interpolation rule,
string interpolation indentation is ignored. Thus, all nodes inside multiline string
got indented the same way the original input which in this case is the `.bad.nix` file.