build: replace flake-compat with flake.lock.nix
This fixes an issue where `nix-build` in the CI wasn't actually building
the project. Because flake-compat exports that weird structure which I
have been trying to argue against (see upstream).

In this mode, flake is only really used for dependency management. As a
sort of niv replacement. The default.nix is the project entry-point, as
we used to do. And dependencies are composed manually.

Start using `nixpkgs` instead of `pkgs` to encourage composition over
inheritance.