Properly handle SIGPIPE
Commands that print their results to stdout should terminate on
SIGPIPE, so that they don't produce error messages when piped into
programs like head(1) or less(1) that might exit before the process is
finished.  This is the default behaviour on Unix, but the Rust runtime
ignores SIGPIPE by default.

For nixpkgs-fmt modes where it makes sense, re-enable the Unix default
behaviour of terminating on SIGPIPE.  Don't do this when formatting
files, because only formatting half the files if stdout was closed
would be a surprising behaviour.